Fix local block file
